[DataMapper] Auto load relationship when using query builder
